requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
feedparser==6.0.11
idna==3.10
iso8601==2.1.0
MarkupSafe==3.0.2
pathlib==1.0.1
psycopg2-binary==2.9.10
reader==3.18
requests==2.32.3
sgmllib3k==1.0.0
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
utils==1.0.2
Werkzeug==3.1.3
